Add feature send no-check-received packet
Previously, every packet was checked if it was received (using sequence
id, ack, and ackbitfield). If a checked packet had timed out, it
would be sent again. Now, a packet can be specified to not include
this check. If such a packet was lost, it would not be resent.

Also, some minor fixes and improvements.